requirements.txt
scipy
imageio
argparse
tensorboardX
blessings
progressbar2
path.py